requirements.txt
Django==2.2.7
django-crispy-forms==1.8.1
django-taggit==1.2.0
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0